+++ src/chart.ts
@@ -1,1 +1,0 @@
-export * from './components/views/UChartView.js';