+++ src/index.ts
@@ -1,5 +1,6 @@
 // components exports
 export * from './components/blocks/UCodeBlock.js';
+export * from './components/blocks/UExtraBlock.js';
 export * from './components/blocks/UFileBlock.js';
 export * from './components/blocks/UMarkedBlock.js';
 export * from './components/blocks/UTextBlock.js';
@@ -12,22 +13,16 @@ export * from './components/references/URefTag.js';
 export * from './components/references/URefCard.js';
 export * from './components/references/URefCardGroup.js';
 
-// view exports
-export * from './components/views/UImagesView.js';
-export * from './components/views/UVideoView.js';
-export * from './components/views/UMapView.js';
-export * from './components/views/UView.js';
-
 // types exports
 export type * from './types/BlockItem';
+export * from './types/Extras';
 export type * from './types/JsonSchema';
 export type * from './types/References';
-export * from './types/Views';
 
 // event types exports
 export type * from './events/SendEvent';
 export type * from './events/StopEvent';
 
 // utilities exports
-export * from './utilities/ViewPromptBuilder.js';
+export * from './utilities/ExtraPromptBuilder.js';
 // export * from './utilities/dom-agent/index.js';